* woa-bk-bcs/master:
  bcs-cluster-manager-0920 (merge request !1438)
  feat: google ca (merge request !1436)
  doc: 文档链接修改 (merge request !1437)

+++ bcs-services/bcs-cluster-manager/internal/store/node/node.go
@@ -35,6 +35,7 @@ const (
 	nodeIPKeyName         = "innerip"
 	nodeClusterIDKey      = "clusterid"
 	nodeTableName         = "node"
+	nodeNameKey           = "nodename"
 	defaultNodeListLength = 5000
 )
 
@@ -151,6 +152,26 @@ func (m *ModelNode) DeleteClusterNode(ctx context.Context, clusterID, nodeID str
 	cond := operator.NewLeafCondition(operator.Eq, operator.M{
 		nodeClusterIDKey: clusterID,
 		nodeIDKeyName:    nodeID,
+	})
+	_, err := m.db.Table(m.tableName).Delete(ctx, cond)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+// DeleteClusterNodeByName delete node by name
+func (m *ModelNode) DeleteClusterNodeByName(ctx context.Context, clusterID, nodeName string) error {
+	if err := m.ensureTable(ctx); err != nil {
+		return err
+	}
+	if len(nodeName) == 0 {
+		return nil
+	}
+
+	cond := operator.NewLeafCondition(operator.Eq, operator.M{
+		nodeClusterIDKey: clusterID,
+		nodeNameKey:      nodeName,
 	})
 	_, err := m.db.Table(m.tableName).Delete(ctx, cond)
 	if err != nil {
@@ -295,6 +316,22 @@ func (m *ModelNode) GetNodeByIP(ctx context.Context, ip string) (*types.Node, er
 	}
 	cond := operator.NewLeafCondition(operator.Eq, operator.M{
 		nodeIPKeyName: ip,
+	})
+	retNode := &types.Node{}
+	if err := m.db.Table(m.tableName).Find(cond).One(ctx, retNode); err != nil {
+		return nil, err
+	}
+	return retNode, nil
+}
+
+// GetNodeByName get node by name
+func (m *ModelNode) GetNodeByName(ctx context.Context, clusterID, name string) (*types.Node, error) {
+	if err := m.ensureTable(ctx); err != nil {
+		return nil, err
+	}
+	cond := operator.NewLeafCondition(operator.Eq, operator.M{
+		nodeClusterIDKey: clusterID,
+		nodeNameKey:      name,
 	})
 	retNode := &types.Node{}
 	if err := m.db.Table(m.tableName).Find(cond).One(ctx, retNode); err != nil {
